First commit after old repo's r33.
Added basic logic for drawing cards and ran StyleCop. Next step is to work out a system for player strategies OR work on the dealer. I think working on a basic game strategy (such as one of the cards they give you at the casino) would be the best bet. Then after the players are doing their thing you can work out how the game actually flows (IE, when cards are dealt, when bets are placed, when cards are picked up, how cards are discarded, and when cards are flipped, etc).


git-svn-id: file:///content/svn/csharp/blackjack@4 dc164686-a85c-4032-854d-e7846e65c01e